feat: combined existing collection authorization handlers/operations, refs AC-1125
